Renaming bpr to fast-boot
