+++ lib/rules/no-timers.js
@@ -1,5 +1,5 @@
 const { getParent } = require('../utils/common');
-const { isGuarded } = require('../utils/bpr');
+const { isGuarded } = require('../utils/fast-boot');
 
 const TIMERS = ['setTimeout', 'setInterval'];
