+++ lib/utils/fast-boot.js
